// src/app/machine/link-type.enum.ts
export enum LinkType {
    ARM = 'ARM',
    BALL_JOINT = 'BALL_JOINT',
    HEXAPOD = 'HEXAPOD',
    LINEAR_JOINT = 'LINEAR_JOINT',
    PIVOT_JOINT = 'PIVOT_JOINT',
    STATIC = 'STATIC',
    TRIPOD = 'TRIPOD',
    UNIVERSAL_JOINT = 'UNIVERSAL_JOINT'
}

export enum LinkState {
    AUXILIARY = 'AUXILIARY',
    DYNAMIC = 'DYNAMIC',
    FROZEN = 'FROZEN',
    STATIC = 'STATIC'
}